+++ CHANGELOG.md
@@ -18,6 +18,9 @@ CHANGELOG
 - Fix cluster creation failure without Internet access when GPU instances and DCV are used.
 - Fix intermittent cluster creation failure caused by eventual consistency issues when head, compute and login nodes have the same security group.
 
+**DEPRECATIONS**
+- The configuration parameter `LoginNodes/Pools/Ssh/KeyName` is no longer supported.
+
 3.14.2
 ------